requirements.txt
pillow==11.2.1
piexif==1.1.3
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
shiboken6==6.9.0

# Optional: Windows toast notifications (non-blocking)
# Uncomment if you want Windows notifications and your Python version is compatible
# win10toast-reborn==1.0.10
